05 - Flex Modal CSS stylesheet
making the modal into a flexbox and giving it a gap and padding. Making the icon a flex shrink so the the icon does not get smashed and a header style to have it centered and boldend all while giving the text under it and the heading a margin to have space so it isn't so cluttered.