06-flex modal modified the styles sheet for a responsive design
I modified the sidebar, lists, and hyperlinks to make them stationary aswell as gave the cards their own set size so that the flex display can evenly arrange them in an order that can mimic the intended outcome. The footer class was made to have a flex display and to align everything in it in the center both horizontally and verdically.